// src/constants/enum/responseCodes.enum.ts
export const enum HTTP_STATUS {
  OK = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  GONE = 410,
  INTERNAL_ERROR = 500,
  VALIDATION_ERROR = 422,
  RATE_LIMITED = 429,
  NOT_IMPLEMENTED = 501
};

export default HTTP_STATUS;